Refactor models

+++ models/Filtered_ad.go
@@ -1,7 +1,7 @@
 package models
 
 type FilteredAd struct {
-	ID           int64
-	AddID        int64
-	UserFilterID int64
+	ID           uint
+	AddID        uint
+	UserFilterID uint
 }
